+++ NEWS.md
@@ -3,3 +3,4 @@
 * Helpful functions to create and work with palettes and colors
 * Some `jak_theme()` tweaks
 * Alpha-diversity metrics
+* Session and package info utilities